Get DESCRIPTION from GitHub Repository

Get DESCRIPTION from GitHub Repository

Usage

get_desc_from_gh(org, repo, ref = "")

Examples

verdepcheck:::get_desc_from_gh("tidyverse", "dplyr")
#> Type: Package
#> Package: dplyr
#> Title: A Grammar of Data Manipulation
#> Version: 1.1.2.9000
#> Authors@R (parsed):
#>     * Hadley Wickham <hadley@posit.co> [aut, cre] (<https://orcid.org/0000-0003-4757-117X>)
#>     * Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>)
#>     * Lionel Henry [aut]
#>     * Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>)
#>     * Davis Vaughan <davis@posit.co> [aut] (<https://orcid.org/0000-0003-4777-038X>)
#>     * Posit Software, PBC [cph, fnd]
#> Description: A fast, consistent tool for working with data frame
#>     like objects, both in memory and out of memory.
#> License: MIT + file LICENSE
#> URL: https://dplyr.tidyverse.org,
#>     https://github.com/tidyverse/dplyr
#> BugReports: https://github.com/tidyverse/dplyr/issues
#> Depends:
#>     R (>= 3.5.0)
#> Imports:
#>     R6,
#>     cli (>= 3.4.0),
#>     generics,
#>     glue (>= 1.3.2),
#>     lifecycle (>= 1.0.3),
#>     magrittr (>= 1.5),
#>     methods,
#>     pillar (>= 1.9.0),
#>     rlang (>= 1.1.0),
#>     tibble (>= 3.2.0),
#>     tidyselect (>= 1.2.0),
#>     utils,
#>     vctrs (>= 0.6.0)
#> Suggests:
#>     DBI,
#>     Lahman,
#>     RMySQL,
#>     RPostgreSQL,
#>     RSQLite,
#>     bench,
#>     broom,
#>     callr,
#>     covr,
#>     dbplyr (>= 2.2.1),
#>     ggplot2,
#>     knitr,
#>     lobstr,
#>     microbenchmark,
#>     nycflights13,
#>     purrr,
#>     rmarkdown,
#>     stringi (>= 1.7.6),
#>     testthat (>= 3.1.5),
#>     tidyr (>= 1.3.0),
#>     withr
#> VignetteBuilder:
#>     knitr
#> Config/Needs/website: tidyverse, shiny, pkgdown,
#>     tidyverse/tidytemplate
#> Config/testthat/edition: 3
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.2.3
verdepcheck:::get_desc_from_gh("tidyverse", "dplyr", "v1.1.0")
#> Type: Package
#> Package: dplyr
#> Title: A Grammar of Data Manipulation
#> Version: 1.1.0
#> Authors@R (parsed):
#>     * Hadley Wickham <hadley@posit.co> [aut, cre] (<https://orcid.org/0000-0003-4757-117X>)
#>     * Romain François [aut] (<https://orcid.org/0000-0002-2444-4226>)
#>     * Lionel Henry [aut]
#>     * Kirill Müller [aut] (<https://orcid.org/0000-0002-1416-3412>)
#>     * Davis Vaughan <davis@posit.co> [aut] (<https://orcid.org/0000-0003-4777-038X>)
#>     * Posit, PBC [cph, fnd]
#> Description: A fast, consistent tool for working with data frame
#>     like objects, both in memory and out of memory.
#> License: MIT + file LICENSE
#> URL: https://dplyr.tidyverse.org,
#>     https://github.com/tidyverse/dplyr
#> BugReports: https://github.com/tidyverse/dplyr/issues
#> Depends:
#>     R (>= 3.4.0)
#> Imports:
#>     R6,
#>     cli (>= 3.4.0),
#>     generics,
#>     glue (>= 1.3.2),
#>     lifecycle (>= 1.0.3),
#>     magrittr (>= 1.5),
#>     methods,
#>     pillar (>= 1.5.1),
#>     rlang (>= 1.0.6),
#>     tibble (>= 2.1.3),
#>     tidyselect (>= 1.2.0),
#>     utils,
#>     vctrs (>= 0.5.2)
#> Suggests:
#>     DBI,
#>     Lahman,
#>     RMySQL,
#>     RPostgreSQL,
#>     RSQLite,
#>     bench,
#>     broom,
#>     callr,
#>     covr,
#>     dbplyr (>= 2.2.1),
#>     ggplot2,
#>     knitr,
#>     lobstr,
#>     microbenchmark,
#>     nycflights13,
#>     purrr,
#>     rmarkdown,
#>     stringi (>= 1.7.6),
#>     testthat (>= 3.1.5),
#>     tidyr (>= 1.3.0),
#>     withr
#> VignetteBuilder:
#>     knitr
#> Config/Needs/website: tidyverse, shiny, pkgdown,
#>     tidyverse/tidytemplate
#> Config/testthat/edition: 3
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.2.3