Temporarily create a valid DESCRIPTION file to a location that will be deleted

The file is deleted after the parent environment where this function was called has exited, when the R session ends or on demand via withr::deferred_run()

Usage

local_description(
  pkg_list = c(pkgdepends = "Import"),
  remotes = c(),
  need_verdepcheck = c(),
  .local_envir = parent.frame()
)

Arguments

pkg_list

(vector) named character vector or list with paired name and type of dependency. It supports versions by using quotes on the key

remotes

(vector) string vector that contains remotes to add to the DESCRIPTION file

need_verdepcheck

(vector) string vector that contains Config/Need/verdepcheck elements to add to the DESCRIPTION file

.local_envir

(envirnoment) The environment to use for scoping.

Examples

verdepcheck:::local_description(
  list(dplyr = "Import"),
  remotes = "tidyverse/dplyr",
  need_verdepcheck = "dplyr=tidyverse/dplyr@v1.1.0"
)
#> [1] "/tmp/Rtmp6b1r1G/DESCRIPTION-27135de936b"