Execute R CMD CHECK on a local package with all dependencies pre-installed using various strategies.

Execute R CMD CHECK on a local package with all dependencies pre-installed using various strategies.

Usage

max_deps_check(
  path,
  config = list(dependencies = .desc_field, library = tempfile()),
  build_args = character(),
  check_args = character(),
  ...
)

release_deps_check(
  path,
  config = list(dependencies = .desc_field, library = tempfile()),
  build_args = character(),
  check_args = character(),
  ...
)

min_deps_check(
  path,
  config = list(dependencies = .desc_field, library = tempfile()),
  build_args = character(),
  check_args = character(),
  ...
)

Arguments

path

(string) path to the package sources

config

(list) configuration options. See pkgdepends::pkg_config for details. If it does not include library then temporary directory is used which simulates clean environment without using any pre-installed packages. If it does not include dependencies then TRUE value is used which means all hard dependencies plus Suggests.

build_args

(string) value passed as build_args argument into rcmdcheck::rcmdcheck()

check_args

(string) value passed as args argument into rcmdcheck::rcmdcheck()

...

other arguments passed to rcmdcheck::rcmdcheck()

Value

a named list with two elements:

"ip" - installation plan object

"check" - returned value from rcmdcheck::rcmdcheck()

strategies

Currently implemented strategies:

max - use the greatest version of dependent packages

release - use released version of dependent packages - use CRAN if possible else if GitHub release is available then use it else fail.

min - use the lowest version of dependent packages incorporating minimal version specification in "Imports" and "Suggests". If no version is specified then the minimal available version is assumed. See get_ref_min for details how the minimal version is determined.

Any modification is done for direct dependencies. Indirect ones are installed as usual.

See also